refactor: create context package
